Fix cpu int -> double

+++ m8s/src/main/java/com/twosigma/m8s/CookJob.java
@@ -9,8 +9,8 @@ import java.util.List;
 
 public class CookJob {
     private String uuid;
-    private int numCpu = 0;
-    private int numMemMB = 0;
+    private double numCpu = 0;
+    private double numMemMB = 0;
 
     public CookJob() {
     }
@@ -19,9 +19,9 @@ public class CookJob {
         return this.uuid;
     }
 
-    public int getNumCpu() { return this.numCpu; }
+    public double getNumCpu() { return this.numCpu; }
 
-    public int getNumMemMB() { return this.numMemMB; }
+    public double getNumMemMB() { return this.numMemMB; }
 
     private static List<CookJob> parseJobArray(JSONArray poolJobs) {
         List<CookJob> jobs = new ArrayList<>();
@@ -35,7 +35,7 @@ public class CookJob {
             for (int j = 0; j < resources.length(); ++j) {
                 JSONObject resource = resources.getJSONObject(j);
                 final String resourceType = resource.getString("resource/type");
-                final int resourceAmount = resource.getInt("resource/amount");
+                final double resourceAmount = resource.getDouble("resource/amount");
                 if (resourceType.equals("resource.type/cpus")) {
                     job.numCpu = resourceAmount;
                 } else if (resourceType.equals("resource.type/mem")) {
